Moved The Types To The Appropriate File
